+++ pioreactor/calibrations/od_calibration.py
@@ -244,7 +244,6 @@ def start_recording_and_diluting(
         unit=get_unit_name(),
         fake_data=is_testing_env(),
         experiment=get_testing_experiment_name(),
-        use_calibration=False,
     ) as od_reader:
 
         def get_voltage_from_adc() -> pt.Voltage: